Replace NewsGrid component with NewsFeed component for improved article display and loading states
